+++ app/Service/DungeonRoute/ThumbnailServiceInterface.php
@@ -25,7 +25,7 @@ interface ThumbnailServiceInterface
         ?int         $viewportHeight = null,
         ?int         $imageWidth = null,
         ?int         $imageHeight = null,
-        ?int         $zoomLevel = null,
+        ?float       $zoomLevel = null,
         ?int         $quality = null,
     ): Collection;
 
@@ -43,7 +43,7 @@ interface ThumbnailServiceInterface
         ?int         $viewportHeight = null,
         ?int         $imageWidth = null,
         ?int         $imageHeight = null,
-        ?int         $zoomLevel = null,
+        ?float       $zoomLevel = null,
         ?int         $quality = null,
     ): ?DungeonRouteThumbnail;
